# docs/openapi.inventory.yml
openapi: 3.0.3
info:
  title: Shiatsu Inventory Backend API
  version: 0.1.0
  description: >
    Contract-first API for Month 2 inventory backend.
    All endpoints require Supabase JWT bearer token except /api/health.
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
...
paths:
...
  /api/productos:
    get:
      tags: [Productos]
      summary: Search products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: sku
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/stock:
    get:
      tags: [Stock]
      summary: Stock query by bodega_id, sku or producto_id
      parameters:
        - name: bodega_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: sku
          in: query
          required: false
          schema:
            type: string
        - name: producto_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stock result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItem'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/movimientos:
    get:
      tags: [Movimientos]
      summary: Query movements
      parameters:
        - name: tipo
          in: query
          schema:
            $ref: '#/components/schemas/MovimientoTipo'
        - name: bodega_id
          in: query
          schema:
            type: string
            format: uuid
        - name: desde
          in: query
          schema:
            type: string
            format: date-time
        - name: hasta
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Movement list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovimientoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Producto:
      type: object
      required: [id, sku, nombre, precio_venta, costo_promedio]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        nombre:
          type: string
        precio_venta:
          type: number
          format: double
        costo_promedio:
          type: number
          format: double
...
    StockItem:
      type: object
      required: [bodega_id, producto_id, sku, nombre, cantidad]
      properties:
        bodega_id:
          type: string
          format: uuid
        producto_id:
          type: string
          format: uuid
        sku:
          type: string
        nombre:
          type: string
        cantidad:
          type: integer
          minimum: 0
        stock_minimo:
          type: integer
          nullable: true
...
    MovimientoTipo:
      type: string
      enum:
        - COMPRA_PROVEEDOR
        - VENTA_POS
        - TRANSFERENCIA_ENVIADA
        - TRANSFERENCIA_RECIBIDA
        - AJUSTE_AUDITORIA
    MovimientoEstado:
      type: string
      enum: [COMPLETADO, EN_TRANSITO]
    MovimientoDetalle:
      type: object
      required: [id, producto_id, cantidad]
      properties:
        id:
          type: string
          format: uuid
        producto_id:
          type: string
          format: uuid
        cantidad:
          type: integer
          minimum: 1
        delta:
          type: integer
          nullable: true
        costo_unitario:
          type: number
          format: double
          nullable: true
    MovimientoResponse:
      type: object
      required: [id, tipo, usuario_id, fecha, detalles]
      properties:
        id:
          type: string
          format: uuid
        tipo:
          $ref: '#/components/schemas/MovimientoTipo'
        estado:
          allOf:
            - $ref: '#/components/schemas/MovimientoEstado'
          nullable: true
        origen_bodega_id:
          type: string
          format: uuid
          nullable: true
        destino_bodega_id:
          type: string
          format: uuid
          nullable: true
        usuario_id:
          type: string
          format: uuid
        proveedor:
          type: string
          nullable: true
        nro_documento:
          type: string
          nullable: true
        motivo:
          type: string
          nullable: true
        evidencia_url:
          type: string
          nullable: true
        fecha:
          type: string
          format: date-time
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/MovimientoDetalle'